kaggle-agent/pyproject.toml
[project]
name = "kaggle-agent"
version = "0.1.0"
description = "Autonomous Kaggle competition research framework inspired by autoresearch + deer-flow"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "lightgbm>=4.5.0",
    "catboost>=1.2.0",
    "optuna>=4.0.0",

    # Data
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "pyarrow>=18.0.0",

    # Deep Learning (optional, for TabNet/NN)
    "torch>=2.5.0",

    # Kaggle API
    "kaggle>=1.6.0",

    # Utilities
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "rich>=13.9.0",
    "click>=8.1.0",
    "tabulate>=0.9.0",

    # Visualization (optional)
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]

[project.scripts]
kaggle-agent = "kaggle_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kaggle_agent"]
